+++ docs/spec/src/index.ts
@@ -1,9 +1,10 @@
 import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
-import { accessSpec } from './access';
 import { accountSpec } from './account';
+import { userSpec } from './user';
+import { userGroupsSpec } from './user-groups';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [...accountSpec, ...accessSpec],
+  integrationSteps: [...accountSpec, ...userGroupsSpec, ...userSpec],
 };