+++ src/steps/index.ts
@@ -1,6 +1,15 @@
 import { accountSteps } from './account';
-import { accessSteps } from './access';
+import { pipelinesSteps } from './pipeline';
+import { projectsSteps } from './project';
+import { userSteps } from './user';
+import { userGroupsSteps } from './user-groups';
 
-const integrationSteps = [...accountSteps, ...accessSteps];
+const integrationSteps = [
+  ...accountSteps,
+  ...userGroupsSteps,
+  ...pipelinesSteps,
+  ...projectsSteps,
+  ...userSteps,
+];
 
 export { integrationSteps };