+++ src/steps/user/index.ts
@@ -1,0 +1,79 @@
+import {
+  createDirectRelationship,
+  Entity,
+  IntegrationStep,
+  IntegrationStepExecutionContext,
+  RelationshipClass,
+} from '@jupiterone/integration-sdk-core';
+
+import { createAPIClient } from '../../client';
+import { IntegrationConfig } from '../../config';
+import {
+  ACCOUNT_ENTITY_KEY,
+  Entities,
+  Relationships,
+  Steps,
+} from '../constants';
+import { createUserEntity } from './converter';
+
+export async function fetchUser({
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const apiClient = createAPIClient(instance.config);
+
+  const user = await apiClient.fetchUser();
+  const userEntity = createUserEntity(user);
+  await jobState.addEntity(userEntity);
+
+  await jobState.iterateEntities(
+    { _type: Entities.USER_GROUP._type },
+    async (userGroupEntity) => {
+      await jobState.addRelationship(
+        createDirectRelationship({
+          _class: RelationshipClass.HAS,
+          from: userGroupEntity,
+          to: userEntity,
+        }),
+      );
+    },
+  );
+}
+
+export async function buildUserAndAccountRelationship({
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const accountEntity = (await jobState.getData(ACCOUNT_ENTITY_KEY)) as Entity;
+
+  await jobState.iterateEntities(
+    { _type: Entities.USER._type },
+    async (userEntity) => {
+      await jobState.addRelationship(
+        createDirectRelationship({
+          _class: RelationshipClass.HAS,
+          from: accountEntity,
+          to: userEntity,
+        }),
+      );
+    },
+  );
+}
+
+export const userSteps: IntegrationStep<IntegrationConfig>[] = [
+  {
+    id: Steps.USERS,
+    name: 'Fetch Users',
+    entities: [Entities.USER],
+    relationships: [Relationships.USER_GROUP_HAS_USER],
+    dependsOn: [Steps.USER_GROUPS],
+    executionHandler: fetchUser,
+  },
+  {
+    id: Steps.BUILD_ACCOUNT_USER_RELATIONSHIPS,
+    name: 'Build Account and User Relationships',
+    entities: [],
+    relationships: [Relationships.ACCOUNT_HAS_USER],
+    dependsOn: [Steps.ACCOUNT, Steps.USERS],
+    executionHandler: buildUserAndAccountRelationship,
+  },
+];